Split up header files into 2.
